+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Extra Icons Change Log
 
-## 2022.1.6 (WIP)
+## 2022.1.6 (2022/06/25)
 * improve accuracy and performance of Angular support.
+* support various Helm files and folders.
 * support Cookie text files.
-* support various Helm files.
+* support `node_modules` folders.
+* add an alternative YAML icon.
+* the icons list in the configuration panel now indicates if you need to restart the IDE to see changes.
+* some code cleanup.
 
 ## 2022.1.5 (2022/06/14)
 * fix #99: Angular support (when it looks for `angular.json` files) may freeze the IDE on some specific projects.